Debug view controller, cleanup

+++ Classes/Libraries/UIColor-Expanded.m
@@ -375,7 +375,7 @@ static NSMutableDictionary *colorNameCache = nil;
 }
 
 - (NSString *)hexStringFromColor {
-	return [NSString stringWithFormat:@"%0.6X", self.rgbHex];
+	return [NSString stringWithFormat:@"%0.6X", (unsigned int) self.rgbHex];
 }
 
 + (UIColor *)colorWithString:(NSString *)stringToConvert {
